pyproject.toml
[tool.poetry]
name = "banking-rtm-pipeline"
version = "0.1.0"
description = "Real-time transaction monitoring pipeline"
authors = ["Sanusi Abdulmalik <sansmal04@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = ">=3.11,<3.12"
confluent-kafka = "^2.3"
pyspark = "^3.5"
delta-spark = "^3.2"
faker = "^24.0"
pandas = "^2.2"
pyarrow = "^17.0"
python-dotenv = "^1.0"
scikit-learn = "^1.4"
xgboost = "^2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
ruff = "^0.4"
mypy = "^1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry-core.masonry.api"